oracle-v3/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
python-dotenv==1.0.0
web3==6.12.0
requests==2.31.0
